Refactor the job/task runner program.

- Change the old term "jobs" to "tasks" to better reflect that they are one-off tasks
- Create a new concept called "jobs", which are like cron jobs
- Move 2 jobs to their own separate source files
- Update setup.sh to reflect the changed filenames
- Renamed the database table "jobs" to "tasks"

+++ server/setup.sh
@@ -12,8 +12,12 @@ mkdir -p /var/www/nivlheim/{db,certs,CA,queue}
 # initialize certificate db
 cd /var/www/nivlheim/db
 touch index.txt
-echo 'unique_subject = no' > index.txt.attr
-echo '100001' > serial
+if [ ! -f index.txt.attr ]; then
+	echo 'unique_subject = no' > index.txt.attr
+fi
+if [ ! -f serial ]; then
+	echo '100001' > serial
+fi
 touch /var/www/nivlheim/rand
 
 # generate a Certificate Authority certificate to sign other certs with
@@ -26,17 +30,19 @@ fi
 
 # generate a SSL certificate as a default for the web server
 cd /var/www/nivlheim
-rm -f default_cert.pem default_key.pem csr
-# key
-openssl genpkey -outform PEM -out default_key.pem -algorithm RSA \
-  -pkeyopt rsa_keygen_bits:4096
-# certificate request
-openssl req -new -key default_key.pem -out csr -days 365 \
-  -subj "/C=NO/ST=Oslo/L=Oslo/O=UiO/OU=USIT/CN=localhost"
-# sign the request
-openssl ca -batch -in csr -cert CA/nivlheimca.crt -keyfile CA/nivlheimca.key \
-  -out default_cert.pem -config /etc/nivlheim/openssl_ca.conf
-rm -f csr
+if [ ! -f default_cert.pem ] || [ ! -f default_key.pem ]; then
+	rm -f default_cert.pem default_key.pem csr
+	# key
+	openssl genpkey -outform PEM -out default_key.pem -algorithm RSA \
+	  -pkeyopt rsa_keygen_bits:4096
+	# certificate request
+	openssl req -new -key default_key.pem -out csr -days 365 \
+	  -subj "/C=NO/ST=Oslo/L=Oslo/O=UiO/OU=USIT/CN=localhost"
+	# sign the request
+	openssl ca -batch -in csr -cert CA/nivlheimca.crt -keyfile CA/nivlheimca.key \
+	  -out default_cert.pem -config /etc/nivlheim/openssl_ca.conf
+	rm -f csr
+fi
 
 # fix permissions
 chgrp -R apache /var/www/nivlheim /var/log/nivlheim
